This commit builds on 08f7fa3 and disables sysmon completely. (#22703)
With this, overnight tests did not display the symptoms of corruption
and no packets on cpu_mnic1 were dropped.
It is not established that sysmon is corrupting the ionic queue
memory, this may be just masking or postponing the problem on the
testbed where this was tested. This is a temp workaround until we find
the real issue.