linux ionic: Backout fw upgrade rework (#22906)
* linux ionic: revert the LIF_READY patch

This was an attempted fix for issues seen with the infamous
fw-upgrade-rework patch.  We're reverting that code so need
to revert this fix as well.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: revert the fw-upgrade-rework

The fw-upgrade-rework patch from PR-22053 needs more stability
work, is not ready for use.  This patch reverts that code but
keeps the rest of the minor bug fixes from those PRs.

Signed-off-by: Shannon Nelson <snelson@pensando.io>