Minimize the output length of driver private stats (#23692)
There's an issue in ESXi itself when printing driver private stats.
No matter how long we claim the private stat length to be, there's
a constraint in kernel to prevent going beyond certain value. As the
time goes, more and more packets have been handled by driver, we will
be facing a situation that some conters are not able to be displayed
in private stats.

Due to another limitation in certification tests, I can't change the
format and abbreviate certain counters (their scripts will be looking
for some names of counters).

What I am doing now is trying my best to minimize the output of driver
private stats.