linux_ionic: station_set and port_init fixes (#23856)
* linux_ionic: leave netdev mac alone after fw-update

When running in a bond setup, or some other potential
configurations, the netdev mac may have been changed from
the default device mac.  Since the userland doesn't know
about the changes going on under the covers in a fw-update
it doesn't know the re-push the mac filter.  The driver
needs to leave the netdev mac filter alone when rebuilding
after the fw-update.

Double commit #23725
Fixes: PS-3151
Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux_ionic: call port_init after fw-upgrade

Since the fw has been re-inited, we need to reset the dma space
for the port information, so let's call port_init again.

Double commit #23745
Signed-off-by: Shannon Nelson <snelson@pensando.io>