Vomero 2 support. (#24422)
Add CPLD ID for Vomero 2.
Vomero's boot sequence is A, B, A, B; it does not use GOLD as a
last resort.  Entry to GOLD FW is via the FORCE_GOLD jumper.
Fix the Macronix CR write to not clobber SR.
Bump to version 4.  Update boards.json.  v4 only mandatory for Vomero cards.