linux_ionic: ifdef the debug stats printing (#24426)
Earlier we ifdef'd out the stats gathering for a few things like
csum and vlan tagging, but we're still printing the values in
the ethtool output, and they come out all 0's.  This patch hides
those values from the ethtool -S output so as to not print bogus data.

If someone wants to run the driver with these enabled, build the
driver with -DIONIC_DEBUG_STATS compiler flag.  One way to do this
is to run the ./build.sh script like this:
	EXTRA_CFLAGS=-DIONIC_DEBUG_STATS ./build.sh

While we're here, we also fix up the apulu version of the privflags test.

Signed-off-by: Shannon Nelson <snelson@pensando.io>