windows: fix OID_GEN_NETWORK_LAYER_ADDRESSES (#24512) (#24535)
* windows: fix OID_GEN_NETWORK_LAYER_ADDRESSES

* windows: init LinkCheckWorker before notifyq

The notifyq becomes active during lif init.  The work event structs need
to be initialized before the notifyq is started.  Otherwise, if an event
is polled, then signaling the event can deref the null pointer list head
when walking the list to wake a potentially waiting thread.

The worker thread is still started in the original place, after all
other initialization is completed.

* windows: no-op OID_GEN_NETWORK_LAYER_ADDRESSES